Refactor dance.py and improve filtering
This commit accomplishes two things. First, it refactors the dance.py
script to use a class in order to better manage the program state.
Thanks to this class, there is no longer a need to pass around various
lists and strings. Second, this commit improves the filtering of
molecules from the database. It selects only molecules with just one
trivalent nitrogen in them and calculates the Wiberg bond order on each.
In the future, molecules will be split into bins based on bond order and
then sorted by size. The smallest molecules in each bin will then be
chosen as the final set of molecules.