Output more data
DANCE was updated to provide more details about each molecule's
trivalent nitrogen and the atoms surrounding it. To do this, a new class
called DanceAnalyzer was created to process molecules coming out of
DanceFilter and save data to files. In addition, DanceProperties was
created (in danceprops) to store molecule properties. DanceFilter was
also modified to calculate these properties, while the overall dance.py
script was updated to take in command line arguments for DanceAnalyzer.
danceutil.py was removed because its attributes fit better in other
files. Finally, corresponding documentation was updated in README. In
the future, DanceAnalyzer will be updated to output plots of the data.