Make PLOTHIST mode more flexible
In order to handle a wider range of CSV files with Wiberg bond orders in
them, PLOTHIST mode was modified to take in a --wiberg-csv flag on the
command line and pass it to DanceWibHist. In DanceWibHist, the column
used for the CSV file was then made to be an instance variable (instead
of the previous hard coding). Finally, various variables across dance.py
and dancewibhist.py were renamed to reflect the increased flexibility.