Rename DanceFilter to DanceGenerator
DanceFilter was renamed to DanceGenerator because the name is more
suitable. DanceGenerator does not really do any filtering; it only
selects trivalent nitrogen molecules. The majority of the actual
"filtering" will be done by SELECT mode later on.

This change was accompanied by several changes in the README, the
invocations and command line help messages in the dance.py script, and
one occurrence of "DanceFilter" in dancesaver.py.