Enable saving of raw Molecules and Properties
DanceSaver now saves to two new files, an OEB (Openeye Binary) file
holding the raw molecules, and a binary file holding DanceProperties
objects. This will ultimately allow us to avoid having to regenerate
molecule properties, particularly the AM1 charges, as those calculations
take a long time to run. However, an update will need to be made to
DanceProperties to save the charge calculations, as they are stored in
an OEAM1Results object, not in the molecule itself.

In addition to DanceSaver itself, the command line arguments in
dance.py were updated, and documentation in README was updated.