Add Saving of Wiberg Bond Orders
Bond orders for each molecule in DanceGenerator are now stored as data
tags for the bonds of a charged copy of the first conformation of the
molecule. This charged copy is then stored in as a data tag in the
original molecule. Once the molecule is saved to the OEB file, the
charged copy and bond orders are saved along with it, helping to avoid
needing to recalculate charges in the future.

Several alternatives were evaluated for this:
- Storing the OEAM1Results holding the bond orders does not work. Using
  Python's pickle fails because the OpenEye objects were generated with
  SWIG. Furthermore, the OEAM1Results somehow does not qualify as an
  "OpenEye object" when being stored in the data tag for a molecule.
- Storing the bond orders as a CSV, JSON, or other data file could work,
  but it would mean an extra file, as well as added complexity.
In short, this option of storing bond orders on the bonds themselves
seems to make the most sense.