Make file-saving easier
Instead of passing in parameters for each file that needs to be saved,
one instead passes in a parameter telling the name of the directory to
save the files in. All the files are then placed in that directory. To
do this, dance.py had to be modified to take in the directory on the
command line and automatically generate the filenames.

README was updated to reflect these changes.