Prep SELECT Mode
Set up the code for using SELECT mode. A class called DanceSelect was
created to handle selecting of the final set of molecules. Code was
added to dance.py to read in molecules and properties and feed them to
DanceSelect, and to read the output of DanceSelect and write it to a
file using DanceSaver. DanceProps had to be updated to support several
operations with the molecules and their properties.

In the future, the molecules and properties may be combined into a
single data structure.