Tune selection
The equality and hash functions for fingerprints were modified so that
objects would fit in the bins they were supposed to. Before, objects
with the same fingerprint were being put into separate bins. The hash
function now simply hashed the data in the fingerprint.

In addition, the default rounding for wiberg bond order was set to 0.05
in order to mkae it easier to get molecules to fall into the same
fingerprint bin.