Minimize data stored in memory
Instead of trying to load all molecules into memory at once during
SELECT mode, only one molecule at a time is now read from the OEB file.
(Note, however, that all the DanceProperties are still stored in memory,
as they are considerably smaller and can still fit). Once the molecule's
fingerprint has been determined, it is appended to an appropriate file
in an output directory, where the file stores all molecules with the
same fingerprint and total wiberg bond order around the atom. Selection
of the smallest molecules from each file will be implemented in the near
future.

Note that DanceSelector was removed because it is no longer necessary.